Fix bug in which clicking on a node may not present the correct details
Previously, if you clicked on node which had children you would be
presented with the rightmost-child's details. This is because I was
reassigning the .data() while drawing subsequences. (I reassigned the
data so that the highlighting logic would work.) I changed the
subsequence drawing strategy to draw from the children up instead of the
parent down.